pyproject.toml
[project]
name = "obsidiansummery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "frontmatter>=3.0.8",
    "openai>=1.60.0",
    "pathlib>=1.0.1",
    "python-dotenv>=1.0.1",
]
